chore: cleanup example
