+++ example/.eslintrc.js
@@ -1,6 +1,6 @@
 module.exports = {
   root: true,
-  extends: ['@react-native', '../config/.eslintrc.js'],
+  extends: ["../config/.eslintrc.js"],
   parserOptions: {
     tsconfigRootDir: __dirname,
     project: true,
