+++ example/src/utils/time.ts
@@ -1,0 +1,6 @@
+export const formatTime = (seconds: number) => {
+  if (isNaN(seconds)) return '--:--';
+  const m = Math.floor(seconds / 60);
+  const s = Math.floor(seconds % 60);
+  return `${m}:${s < 10 ? '0' : ''}${s}`;
+};
