+++ packages/drm-plugin/package.json
@@ -64,19 +64,13 @@
     "registry": "https://registry.npmjs.org/"
   },
   "devDependencies": {
-    "@eslint/compat": "^1.2.7",
-    "@eslint/eslintrc": "^3.3.0",
-    "@eslint/js": "^9.22.0",
     "@react-native/babel-preset": "0.79.2",
-    "@react-native/eslint-config": "^0.78.0",
     "@release-it/conventional-changelog": "^9.0.2",
-    "@types/jest": "^29.5.5",
     "@types/react": "^18.3.0",
     "del-cli": "^5.1.0",
-    "eslint": "^9.22.0",
-    "eslint-config-prettier": "^10.1.1",
-    "eslint-plugin-prettier": "^5.2.3",
-    "jest": "^29.7.0",
+    "eslint": "^8.51.0",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-plugin-prettier": "^5.0.1",
     "nitro-codegen": "^0.27.6",
     "prettier": "^3.0.3",
     "react": "18.3.1",
@@ -84,7 +78,6 @@
     "react-native-builder-bob": "^0.40.13",
     "react-native-nitro-modules": "^0.27.6",
     "release-it": "^17.10.0",
-    "turbo": "^1.10.7",
     "typescript": "^5.8.3",
     "react-native-video": "*"
   },
